wip: assigning and inserting permissions
Create permission_assignment_activities table to create a locus of activity.
This allows the application to stream write "assign permission" requests to a
single table, permission_assignment_activities. The table is designed to
capture everything present in a typical request as well as gather information
needed for validating the change.  Since permission_assignment_activities is a
table, and not a view, inserts are trivial.